Update: reorganize actions
